clean up and move saveschedule logic to saveschedule button component
